HAPP-1399: Copy email to clipboard when there is no client
