+++ app/src/main/res/values/strings.xml
@@ -8,6 +8,7 @@
     <string name="yes">Yes</string>
     <string name="no">No</string>
     <string name="not_available">Not available</string>
+    <string name="email_copied_to_clipboard">Email address copied to clipboard</string>
 
     <!--Informational screen for camera permission-->
     <string name="privacy_policy">Privacy Policy</string>
